requirements.txt
Django==3.1.*
djangorestframework==3.12.*
celery==5.0.*
requests==2.25.*
redis==3.5.*
django-environ==0.4.*
psycopg2==2.8.*